Adding popular DICOM-Extensions
Registering other popular DICOM-Extensions for this wonderful Dicomviewer!
Opening of the viewer by clicking on the various files worked with all file extensions in Release 0.0.3.
In current release 1.0.2 the viewer window unfortunately only opens if extension is "dcm". Is there any way to enable it for the other extensions, too?